mini-system
This is my mini-system when you can create a user and transfer money to another client, i can't implement a bank acquiring cause i don't have a big experience in programming exactly in PHP. So that's all what i can do.  Before to start and run this project , you must download this project and send this in your direction:XAMPP/htdocs. then run a PHP and Apache  in your XAMPP. Then u must goingo to your 127.0.0.1:8000/admin direction to see this project.
Admin login:admin
Admin password:admin123
Thanks for this possibility and task cause this is a best experience for me, because i am just starter IT programmer and student, so i'm sorry for i couldn't finish this project like in Full Task with bank acquriing, but if u choose me to Intern I am ready to hard study and work for my future.
P.S. Yerhigit Uzak